+++ CHANGES.rst
@@ -10,6 +10,7 @@ in progress
 - Delegate MQTT message processing to separate thread
 - Run “CREATE DATABASE” only once to improve performance
 - Accept timestamp field ``time`` from sensor readings
+- Data acquisition channel using CSV over HTTP
 
 
 .. _kotori-0.10.10: